services/web/requirements.txt
Flask==2.0.2
Flask-WTF==1.0.0
flask-mongoengine==1.0.0
Flask-Login==0.5.0
clickhouse-driver==0.2.7
Flask-Mail==0.9.1
gunicorn==20.1.0
paho-mqtt==1.6.1
python-daemon==2.3.0
Werkzeug==2.0.2
itsdangerous==2.0.1
WTForms==3.0.0
mongoengine==0.23.1
pyjwt==2.8.0
peewee==3.17.1
mysqlclient==2.2.4

# cffi==1.14.5
# click==7.1.2
# cryptography==3.4.6
# dnspython==2.1.0
# docutils==0.16
# email-validator==1.1.2
# idna==3.1
# Jinja2==2.11.3
# lockfile==0.12.2
# MarkupSafe==1.1.1
# pycparser==2.20
# pymongo==3.11.3
# pytz==2021.1
# tzlocal==2.1
